Move database files to models folder
