+++ server.js
@@ -3,43 +3,20 @@ const express = require("express");
 const cors = require("cors");
 const mongoose = require('mongoose');
 
+// Get models
+const Food = require('./models/food');
+const User = require('./models/user');
+
 mongoose.connect(process.env.DB_CONNECTION, {useNewUrlParser: true, useUnifiedTopology: true});
 console.log(process.env.DB_CONNECTION);
 
-const FoodSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: [true, "needs a name"]
-  },
-  time: {
-    type: Number,
-    required: [true, "number needed"]
-  }
-});
 
-const UserSchema = new mongoose.Schema({
-  username: {
-    type: String,
-    required: [true, "username needed"]
-  },
-  password: {
-    type: String,
-    required: [true, "password needed"]
-  },
-  email: {
-    type: String,
-    required: [true, "email needed"]
-
-  },
-});
-
-const Food = mongoose.model("food", FoodSchema);
 
 // Testing save
-// const food = new Food({
-// name: "Almonds",
-//   time: 10
-// }).save().then( () => console.log("Food saved"));
+ const food = new Food({
+ name: "Almonds",
+   time: 10
+ }).save().then( () => console.log("Food saved"));
 
 const app = express();
